backend/requirements.txt
Django>=2.2.6,<2.3.0
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.7.5,<2.8.0

flake8>=3.6.0,<3.7.0

